vertices

+++ project/engine/3d/Line3D.h
@@ -69,7 +69,7 @@ private://メンバ変数
 	//Line用の頂点リソースを作る
 	std::list<LineData> vertices;
 	std::list<uint32_t> indices;
-	uint32_t preVerticesSize = 0;
+	size_t preVerticesSize = 0;
 	//今のインスタンス数
 	uint32_t kNumVertex_ = 0;
 	uint32_t kNumIndex_ = 0;
